[[ Feature ]] Implement 'try' function
This patch implements a new function 'try'. The try function
takes one or two parameters. The first parameter is the
expression to try and evaluate, and the second (optional)
parameter is the failure expression.

The first expression is executed and if an error is thrown
during evaluation, the second expression (if specified) is
evaluated. If there is no second expression then empty is
returned, and no error propagates.

The second expression may contain the literal 'error',
which will evaluate to the error which was thrown during
evaluation of the first expression.

If evaluating the second expression throws an error, then
that error is propagated.