[[ Feature ]] Implement meta indicies
This patch implements three 'meta' indicies 'first', 'last'
and 'next'. These literals can appear in an array index
expression.

The first and last literals map to the first index (1)
and the last index (the number of elements) of a sequence
(array). If they are used on a non-sequence array value an
error is thrown.

The next literal maps to the element after the last of a
sequence (array) - equivalent to using the number of elements
+ 1.