[[ Foundation ]] Minimise size of core value structs
This patch minimises the size of the libfoundatino core value
structures. Specifically it shifts the order of some fields in
__MCString and __MCData resulting in a reduction in size on
64-bit, and moves the __MCName hash field into the __MCValue
flags word (at the expense of losing 5 bits of hash).

The effect on this on memory utilization will vary depending
on platform and architecture - most critically, the effacy of
the implementatin of malloc for small object sizes.

However, on 64-bit Mac, it means that __MCName and __MCString
will both fit into a 32-byte (no overhead) malloc block, as
opposed to a 48-byte (no overhead) malloc block as occurred
previously.