[[ Perf ]] Implement expression attributes
This patch implements the notion of expression attributes on
MCExpression AST nodes.

The MCExpression base-class has been augmented with a new
virtual method 'getattrs()', which returns an MCExpressionAttrs
struct.

The MCExpressionAttrs currently only has an 'IsConstant' member,
but could be expanded in the future to add more information about
the AST node.

The MCConstant and MCLiteral expression nodes return attrs with
IsConstant true.