[[ Perf ]] Implement 'tagged' numbers
This patch implements 'tagged' numbers which allow 'small'
integers to be stored in the value-ref pointer itself, rather
than require a full __MCNumber struct to be allocated.

Currently, the bottom-most bit of the pointer field is used to
determine whether the number has its value stored in the pointer
field or not.

The CreateWithInteger number constructor will always create a
tagged number if the provided integer will fit in the remaining
bits of the pointer.

A CreateWithAlignedPointer constructor and associated
FetchAsAlignedPointer method has been added. These allow an
(unmanaged) pointer which is aligned with default pointer
alignment (e.g. typically 8 byte aligned on 32-bit and 16
byte aligned on 64-bit) to be represented as a MCNumberRef with
no overhead. Critically this allows the MCValueRef containers
to hold non-valueref pointers - allowing them to be used to
store pointer-based values which aren't valuerefs (as long as
the lifetime of such things are managed externally).