[[ Bug 12419 ]] Ensure Android ScrollViews don't cancel touch events
This patch implements the `onInterceptTouchEvent` method in the vertical and
horizontal scroll views so that they do not cancel the touch events. Without
this patch the cancellation of the touch events causes a touch down to be unpaired
with an up or release event. Not that iOS Scrollers have a `canCancelTouches`
property but android scrollers do not.

(cherry picked from commit 63bc5f4ab9fc1c047db9d732f249b98b9a597273)